sources/requirements.txt
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0